+++ R/timeManip.R
@@ -4,45 +4,45 @@
 #'
 #' @param fromPeriod dateChr "YYYYmmddHH..."
 #' @param toPeriod dateChr "YYYYmmddHH..."
-#' @param timeResolution choice between "daily","hourly","minute","second",...
+#' @param timeResolution "daily","hourly","minute","second",...
+#' @param precision "daily","hourly","minute","second",...
 #' @param Timeresinsec integer
 #' @param nbStep integer
 #' @param seqPeriod POSIXct timeserie
 #' @keywords timeManip
 #' @export
 #' @examples
-#' timeManip(fromPeriod="2013060205",toPeriod="2013060205",timeResolution="hourly")
+#' timeManip(fromPeriod="2013060205",toPeriod="2013060205",timeResolution="hourly",precision="hourly")
 #'
 
-timeManip  <- function(fromPeriod,toPeriod,timeResolution,Timeresinsec=NULL,nbStep=NULL,seqPeriod=NULL){
-   myenv <- environment()
+timeManip  <- function(fromPeriod,toPeriod,timeResolution,precision,Timeresinsec=NULL,nbStep=NULL,seqPeriod=NULL){
 
-   if ((is.null(Timeresinsec)) && (is.null(nbStep)) && (is.null(seqPeriod))) {
-     myfromPeriod <- fromPeriod
-     mytoPeriod <- toPeriod
-     mytimeResolution <- timeResolution
-     myTimeresinsec <- insec(timeResolution=timeResolution)
-     tmp <- dateTimeSerie(mytimeResolution,myfromPeriod,mytoPeriod)
-     mynbStep <- tmp$nbStep
-     myseqPeriod <- tmp$seqPeriod
-   } else {
-     myfromPeriod <- fromPeriod
-     mytoPeriod <- toPeriod
-     mytimeResolution <- timeResolution
-     myTimeresinsec <- Timeresinsec
-     mynbStep <- nbStep
-     myseqPeriod <- seqPeriod
-   }
+   timeManip_object(fromPeriod,toPeriod,timeResolution,precision,Timeresinsec,nbStep,seqPeriod)
 
-   timeManipL = list(
-       fromPeriod     = function(){return(get("myfromPeriod",envir=myenv))},
-       toPeriod       = function(){return(get("mytoPeriod",envir=myenv))},
-       timeResolution = function(){return(get("mytimeResolution",envir=myenv))},
-       Timeresinsec   = function(){return(get("myTimeresinsec",envir=myenv))},
-       nbStep         = function(){return(get("mynbStep",envir=myenv))},
-       seqPeriod      = function(){return(get("myseqPeriod",envir=myenv))})
+}
 
-    assign('this',timeManipL,envir=myenv )
-    class(timeManipL) <- append("timeManip","list")
-    return(timeManipL)
+
+timeManip_object <- function(fromPeriod,toPeriod,timeResolution,precision,Timeresinsec,nbStep,seqPeriod){
+
+  if ((is.null(Timeresinsec)) && (is.null(nbStep)) && (is.null(seqPeriod))) {
+    Timeresinsec <<- insec(timeResolution=timeResolution)
+    tmp <- timeserie(timeResolution,fromPeriod,toPeriod,precision)
+    nbStep <<- tmp$nbStep
+    seqPeriod <<- tmp$seqPeriod
+  }
+
+  object <- local({
+
+     fromPeriod     = function(){return(fromPeriod)}
+     toPeriod       = function(){return(toPeriod)}
+     timeResolution = function(){return(timeResolution)}
+     Timeresinsec   = function(){return(Timeresinsec)}
+     nbStep         = function(){return(nbStep)}
+     seqPeriod      = function(){return(seqPeriod)}
+
+     environment()
+   })
+   lockEnvironment(object, TRUE)
+   structure(object, class=c("timeManip", class(object)))
+
 }